+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -261,6 +261,8 @@ public class SampleSQLBuilder {
 						).put(
 							"dLDataFactory", new DLDataFactory()
 						).put(
+							"fragmentDataFactory", new FragmentDataFactory()
+						).put(
 							"journalDataFactory", new JournalDataFactory()
 						).put(
 							"messageBoardDataFactory",
